% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC.R
\name{CC}
\alias{CC}
\title{CC function run CloudCompare software}
\usage{
CC(cc_syntax)
}
\arguments{
\item{cc_syntax}{- CloudCompare Command Line Mode syntax, character, output of cloudcompare package functions}
}
\value{
- nothing
}
\description{
CC function run CloudCompare software
}
\examples{
\dontrun{
#' pc_files <- list.files('C://', pattern='.las$',full.names=T)
CC(merge_clouds(files= pc_files))
}


}
