% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cc.R
\name{extract_cc}
\alias{extract_cc}
\title{Extract Conected Components}
\usage{
extract_cc(
  file,
  octree_level = 8,
  nop = 50,
  global_shift = F,
  global_shift_type = "AUTO",
  c_export_fmt = "LAS",
  c_ext = "las",
  silent = T,
  no_timestamp = T
)
}
\arguments{
\item{file}{a}

\item{octree_level}{a}

\item{nop}{a}

\item{global_shift}{a}

\item{global_shift_type}{a}

\item{c_export_fmt}{a}

\item{c_ext}{a}

\item{silent}{a}

\item{no_timestamp}{a}
}
\value{

}
\description{
Extract Conected Components
}
\examples{
\dontrun{
#' pc_files <- list.files('C://', pattern='.las$',full.names=T)
CC(merge_clouds(files= pc_files))
}


}
