% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.R
\name{feature}
\alias{feature}
\title{Compute Geometric Feature}
\usage{
feature(
  file,
  type = "SUM_OF_EIGENVALUES",
  radius = 0.01,
  output_dir = "C:\\\\output.las",
  global_shift = F,
  global_shift_type = "AUTO",
  filter_sf = F,
  filter_value = c(0, 1),
  c_export_fmt = "LAS",
  c_ext = "las",
  silent = T,
  no_timestamp = T
)
}
\arguments{
\item{file}{a}

\item{type}{a}

\item{radius}{a}

\item{output_dir}{a}

\item{global_shift}{a}

\item{global_shift_type}{a}

\item{filter_sf}{a}

\item{filter_value}{a}

\item{c_export_fmt}{a}

\item{c_ext}{a}

\item{silent}{a}

\item{no_timestamp}{a}
}
\value{

}
\description{
Compute Geometric Feature
}
\examples{
\dontrun{
#' pc_files <- list.files('C://', pattern='.las$',full.names=T)
CC(merge_clouds(files= pc_files))
}



}
