% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_clouds.R
\name{merge_clouds}
\alias{merge_clouds}
\title{Merge Clouds}
\usage{
merge_clouds(
  files,
  global_shift = F,
  global_shift_type = "AUTO",
  c_export_fmt = "LAS",
  c_ext = "las",
  silent = T,
  no_timestamp = T
)
}
\arguments{
\item{files}{a}

\item{global_shift}{a}

\item{global_shift_type}{a}

\item{c_export_fmt}{a}

\item{c_ext}{a}

\item{silent}{a}

\item{no_timestamp}{a}
}
\value{

}
\description{
Merge Clouds
}
\examples{
\dontrun{
#' pc_files <- list.files('C://', pattern='.las$',full.names=T)
CC(merge_clouds(files= pc_files))
}
}
