% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_CCdir.R
\name{set_CCdir}
\alias{set_CCdir}
\title{Set CloudCompare directory}
\usage{
set_CCdir(CCdir = "C:\\\\Program Files\\\\CloudCompare\\\\CloudCompare.exe")
}
\arguments{
\item{CCdir}{- character, directory to CloudCompare.exe file}
}
\value{

}
\description{
Set CloudCompare directory
}
\examples{
\dontrun{
#' pc_files <- list.files('C://', pattern='.las$',full.names=T)
CC(merge_clouds(files= pc_files))
}


}
